refactor(icon): cleanup based on pr review
simplier concat; better dist filenames; move icon transoformation to js folder; general cleanup
