+++ packages/icon/README.md
@@ -11,6 +11,6 @@ This is currently not working as automatically as we'd like. If there is another
 1. Put in the `src/svg` directory
 1. Add `role="img" and aria-label="<iconId here> icon"` to the root `svg` element of each icon
 1. Manually refactor the svg to not use xlink/defs.
-1. Run `npm run build` in the `packages/icon` directory (new `*.svg.js` files will generate new icon ids automatically)
+1. Run `npm run build` in the `packages/icon` directory (new `*.svg.dist.js` files will generate new icon ids automatically)
 1. Manually spot check the look of the new icons in the reference site at `/components/icon`
 1. Update the snapshot test
